+++ server/src/main/java/replicant/server/ServerConstants.java
@@ -20,23 +20,25 @@ public final class ServerConstants {
      */
     public static final String REQUEST_ID_KEY = "RequestID";
     /**
-     * Key used to retrieve a flag whether the request produced a changeset relevant for the initiating session.
+     * Key used to retrieve a flag indicating whether the request produced a Change Set relevant to the initiating
+     * Replicant Session.
      * Used to pass data from the EJB to the servlet.
      */
     public static final String REQUEST_COMPLETE_KEY = "RequestComplete";
     /**
-     * Key used to retrieve the encoded response for the request relevant for the initiating session.
+     * Key used to retrieve the encoded response for the request relevant to the initiating Replicant Session.
      * Used to pass data from the EJB to the servlet.
      */
     public static final String REQUEST_RESPONSE_KEY = "RequestResponse";
     /**
      * Key used to flag that a Cacheable Dataset Change Set or use-dataset-cache-entry message has been queued.
-     * This means there should be no other changes in the session Change Set and it should be marked as not required.
+     * This means there should be no other changes in the Initiating Session Change Set and it should be marked as not
+     * required.
      */
     public static final String DATASET_CACHE_ENTRY_HANDLED_KEY = "DatasetCacheEntryHandled";
     /**
      * Key used to flag that an action is subscription.
-     * This means there are ZERO changes in session changeset.
+     * This means there are zero changes in the Initiating Session Change Set.
      */
     public static final String SUBSCRIPTION_REQUEST_KEY = "SubscriptionRequest";